Provide an implementation to use when cgo is not available
This does not yet work correctly, see http://stackoverflow.com/questions/38922080/how-can-i-fallback-to-a-go-implementation-when-cgo-is-not-available-during-build
